chore(wallet): pr cleanup
