+++ src/components/Slices/SliceDetails.jsx
@@ -11,7 +11,6 @@ import {
   AppBar,
   makeStyles,
 } from "@material-ui/core";
-import { useSelector } from "react-redux";
 
 import UTXOSet from "../ScriptExplorer/UTXOSet";
 import MultisigDetails from "../MultisigDetails";
@@ -19,7 +18,6 @@ import ImportAddressesButton from "../ImportAddressesButton";
 import ConfirmAddress from "./ConfirmAddress";
 
 import { slicePropTypes, clientPropTypes } from "../../proptypes";
-import { getWalletConfig } from "../../selectors/wallet";
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -75,7 +73,6 @@ TabPanel.defaultProps = {
 
 const SliceDetails = ({ slice, client, network }) => {
   const [tabIndex, setTabIndex] = React.useState("0");
-  const walletConfig = useSelector(getWalletConfig);
   const classes = useStyles();
 
   const handleChange = (_e, newIndex) => setTabIndex(newIndex);
@@ -98,11 +95,7 @@ const SliceDetails = ({ slice, client, network }) => {
           <Typography variant="caption">
             (not all device variants supported)
           </Typography>
-          <ConfirmAddress
-            slice={slice}
-            network={network}
-            walletConfig={walletConfig}
-          />
+          <ConfirmAddress slice={slice} network={network} />
         </div>
       ),
     },
