+++ src/selectors/wallet.js
@@ -14,6 +14,8 @@ const getTotalSigners = (state) => state.settings.totalSigners;
 const getRequiredSigners = (state) => state.settings.requiredSigners;
 const getStartingAddressIndex = (state) => state.settings.startingAddressIndex;
 const getWalletName = (state) => state.wallet.common.walletName;
+const getExtendedPublicKeyImporters = (state) =>
+  state.quorum.extendedPublicKeyImporters;
 const getWalletLedgerPolicyHmacs = (state) =>
   state.wallet.common.ledgerPolicyHmacs;
 const getClientDetails = (state) => {
@@ -264,4 +266,19 @@ export const getWalletDetailsText = createSelector(
 export const getWalletConfig = createSelector(
   [getWalletDetailsText],
   JSON.parse
+);
+
+export const getHmacsWithName = createSelector(
+  getExtendedPublicKeyImporters,
+  getWalletLedgerPolicyHmacs,
+  (extendedPublicKeys, policyHmacs) => {
+    return Object.values(extendedPublicKeys)
+      .map((importer) => {
+        const policyHmac = policyHmacs.find(
+          (hmac) => hmac.xfp === importer.rootXfp
+        )?.policyHmac;
+        return { policyHmac, name: importer.name };
+      })
+      .filter((registration) => registration.policyHmac);
+  }
 );
